Split failure message out of conflict resolver
This will be required to separate the checking for rejected versions
from conflict resolution.
